MCRP test & test folder organization
