+++ tests/PipelineNamingTest.cpp
@@ -1,5 +1,5 @@
 /*
- * PerformanceTest.cpp
+ * PipelineNamingTest.cpp
  *
  *  Created on: Feb 8, 2021
  *      Author: toky
